Add script for resetting Docker

+++ scripts/reset-docker.sh
@@ -2,7 +2,11 @@
 
 docker kill $(docker ps -q)
 docker rm $(docker ps -aq)
-docker image prune -af
-docker volume prune -f
-docker builder prune -af
-docker system prune --volumes -af
+
+if [ "$1" == "-a" ]
+then
+    docker image prune -af
+    docker volume prune -f
+    docker builder prune -af
+    docker system prune --volumes -af
+fi
